% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_CI.R
\name{coef_CI}
\alias{coef_CI}
\title{Confidence Intervals and Estimates of Each Regression Coefficient}
\usage{
coef_CI(lrbs, alpha = 0.05)
}
\arguments{
\item{lrbs}{A linear_reg_bs or linear_reg_bs_par object containing BLB regression
coefficient estimates.}

\item{alpha}{The significance level. Default value is 0.05.}
}
\value{
The overall confidence interval for each regression coefficient, along with
its overall estimate.
}
\description{
This function determines overall confidence intervals and estimates for each
regression coefficient.
(EXPAND ON THIS)
}
