% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linear_reg_bs.R
\name{linear_reg_bs}
\alias{linear_reg_bs}
\title{Linear Regression Using Bag of Little Bootstraps}
\usage{
linear_reg_bs(x, y, s = 10, r = 1000)
}
\arguments{
\item{x}{A dataframe of the explanatory variables of all observations.}

\item{y}{A numeric vector of the response variable of all observations.}

\item{s}{The number of subsamples to split the data into. Default value is 10.}

\item{r}{The number of bootstrap samples to generate from each subsample. Default
value is 1000.}
}
\value{
bootstrap_coefficient_estimates: The estimates of all regression coefficients
for each bootstrap sample. This list has an element for each subsample, and each
element stores the estimates for each bootstrap sample in a matrix.

bootstrap_s2_estimates: The estimates of sigma-squared for each bootstrap
sample. This list has an element for each subsample, and each element stores the
estimates for each bootstrap sample in a vector.
}
\description{
This function fits a linear regression model using BLB.
(EXPAND ON THIS)
}
