% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PI_par.R
\name{PI_par}
\alias{PI_par}
\title{Prediction Intervals and Estimates for New Data - Parallelized}
\usage{
PI_par(lrbs, x, alpha = 0.05)
}
\arguments{
\item{lrbs}{A linear_reg_bs or linear_reg_bs_par object containing BLB regression
coefficient estimates.}

\item{x}{A dataframe of the explanatory variables of unseen observations.}

\item{alpha}{The significance level. Default value is 0.05.}
}
\value{
The prediction intervals and estimates for the response variable of each
unseen observation.
}
\description{
This function determines prediction intervals and estimates for new data using the
BLB regression coefficient estimates and parallelization.
(EXPAND ON THIS)
}
