% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/s2_CI_par.R
\name{s2_CI_par}
\alias{s2_CI_par}
\title{Confidence Intervals and Estimates of Sigma-Squared - Parallelized}
\usage{
s2_CI_par(lrbs, alpha = 0.05)
}
\arguments{
\item{lrbs}{A linear_reg_bs or linear_reg_bs_par object containing BLB sigma-squared
estimates.}

\item{alpha}{The significance level. Default value is 0.05.}
}
\value{
The overall confidence interval for sigma-squared, along with its overall
estimate.
}
\description{
This function determines an overall confidence interval and estimate for
sigma-squared using parallelization.
(EXPAND ON THIS)
}
